Fixes #24227 - Add permissions info to models API
During the process of moving Hardware Models to a React page
the current API was missing significant details that are rendered
in the Backend before the page is sent to the user.

For example, whether the user is allowed to edit a model or not

This information should be moved to the API because in the new
page, once the page is rendered by Rails any communication is
going to be done through the API and by that Rails won't
render the page or part of it again.

Signed-off-by: Boaz Shuster <boaz.shuster.github@gmail.com>